--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -30,7 +30,7 @@ func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 type users [100_000]User
 
 func getUsers(r io.Reader) (result users, err error) {
-	content, err := io.ReadAll(r)
+	content, err := io.ReadAll(r) // порционное чтение?
 	if err != nil {
 		return
 	}
@@ -38,7 +38,7 @@ func getUsers(r io.Reader) (result users, err error) {
 	lines := strings.Split(string(content), "\n")
 	for i, line := range lines {
 		var user User
-		if err = json.Unmarshal([]byte(line), &user); err != nil {
+		if err = json.Unmarshal([]byte(line), &user); err != nil { // try codegen?
 			return
 		}
 		result[i] = user
@@ -52,9 +52,11 @@ func countDomains(u users, domain string) (DomainStat, error) {
 	for _, user := range u {
 		matched := strings.Contains(user.Email, "."+domain)
 		if matched {
-			num := result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])]
+			emailPart := strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])
+
+			num := result[emailPart]
 			num++
-			result[strings.ToLower(strings.SplitN(user.Email, "@", 2)[1])] = num
+			result[emailPart] = num
 		}
 	}
 	return result, nil