--- a/hw10_program_optimization/stats.go
+++ b/hw10_program_optimization/stats.go
@@ -18,7 +18,7 @@ type DomainStat map[string]int
 func GetDomainStat(r io.Reader, domain string) (DomainStat, error) {
 	result := make(DomainStat)
 	scanner := bufio.NewScanner(r)
-	var json = jsoniter.ConfigCompatibleWithStandardLibrary
+	var json = jsoniter.ConfigFastest
 
 	for scanner.Scan() {
 		var user UserEmail