--- a/hw10_program_optimization/stats_test.go
+++ b/hw10_program_optimization/stats_test.go
@@ -1,3 +1,4 @@
+//go:build !bench
 // +build !bench
 
 package hw10programoptimization
@@ -35,5 +36,45 @@ func TestGetDomainStat(t *testing.T) {
 		result, err := GetDomainStat(bytes.NewBufferString(data), "unknown")
 		require.NoError(t, err)
 		require.Equal(t, DomainStat{}, result)
+	})
+
+	t.Run("empty file", func(t *testing.T) {
+		data := ``
+
+		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
+
+		require.NoError(t, err)
+		require.Len(t, result, 0)
+	})
+
+	t.Run("empty jsons", func(t *testing.T) {
+		data := `{}
+{}`
+
+		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
+
+		require.NoError(t, err)
+		require.Len(t, result, 0)
+	})
+
+	t.Run("no emails", func(t *testing.T) {
+		data := `{"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}
+{"Id":2,"Name":"Jesse Vasquez","Username":"qRichardson","Phone":"9-373-949-64-00","Password":"SiZLeNSGn","Address":"Fulton Hill 80"}`
+
+		result, err := GetDomainStat(bytes.NewBufferString(data), "com")
+
+		require.NoError(t, err)
+		require.Len(t, result, 0)
+	})
+}
+
+func TestGetDomainStatErrors(t *testing.T) {
+	t.Run("corrupted json", func(t *testing.T) {
+		data := `"Id":1,"Name":"Howard Mendoza","Username":"0Oliver","Email":"aliquid_qui_ea@Browsedrive.gov","Phone":"6-866-899-36-79","Password":"InAQJvsq","Address":"Blackbird Place 25"}`
+
+		result, err := GetDomainStat(bytes.NewBufferString(data), "gov")
+
+		require.Nil(t, result)
+		require.ErrorContains(t, err, "get users error")
 	})
 }